API Reference

API Reference

Core Functions

ADCME.add_collection — Method.

add_collection(name::String, v::PyObject)

Adds v to the collection with name name. If name does not exist, a new one is created.

source

ADCME.add_collection — Method.

add_collection(name::String, vs::PyObject...)

Adds operators vs to the collection with name name. If name does not exist, a new one is created.

source

ADCME.get_collection — Function.

get_collection(name::Union{String, Missing})

Returns the collection with name name. If name is missing, returns all the trainable variables.

source

ADCME.has_gpu — Method.

has_gpu()

Checks if GPU is available.

Note

ADCME will use GPU automatically if GPU is available. To disable GPU, set the environment variable ENV["CUDA_VISIBLE_DEVICES"]="" before importing ADCME

source

ADCME.if_else — Method.

if_else(condition::Union{PyObject,Array,Bool}, fn1, fn2, args...;kwargs...)

If condition is a scalar boolean, it outputs fn1 or fn2 (a function with no input argument or a tensor) based on whether condition is true or false.

If condition is a boolean array, if returns condition .* fn1 + (1 - condition) .* fn2

source

ADCME.reset_default_graph — Method.

reset_default_graph()

Resets the graph by removing all the operators.

source

ADCME.stop_gradient — Method.

stop_gradient(o::PyObject, args...;kwargs...)

Disconnects o from gradients backpropagation.

source

ADCME.tensor — Method.

tensor(s::String)

Returns the tensor with name s. See tensorname.

source

ADCME.tensorname — Method.

tensorname(o::PyObject)

Returns the name of the tensor. See tensor.

source

ADCME.run_profile — Method.

run_profile(args...;kwargs...)

Runs the session with tracing information.

source

ADCME.save_profile — Function.

save_profile(filename::String="default_timeline.json")

Save the timeline information to file filename.

Open Chrome and navigate to chrome://tracing

Load the timeline file

source

Base.bind — Method.

bind(op::PyObject, ops...)

Adding operations ops to the dependencies of op. The function is useful when we want to execute ops but ops is not in the dependency of the final output. For example, if we want to print i each time i is evaluated

i = constant(1.0)
op = tf.print(i)
i = bind(i, op)

source

Variables

ADCME.TensorArray — Function.

TensorArray(size_::Int64=0, args...;kwargs...)

Constructs a tensor array for while_loop.

source

ADCME.Variable — Method.

Variable(initial_value;kwargs...)

Constructs a ref tensor from value.

source

ADCME.constant — Method.

constant(value; kwargs...)

Constructs a non-trainable tensor from value.

source

ADCME.convert_to_tensor — Method.

convert_to_tensor(o::Union{PyObject, Number, Array{T}, Missing, Nothing}; dtype::Union{Type, Missing}=missing) where T<:Number

Converts the input o to tensor. If o is already a tensor and dtype (if provided) is the same as that of o, the operator does nothing. Otherwise, convert_to_tensor converts the numerical array to a constant tensor or casts the data type.

source

ADCME.gradient_checkpointing — Function.

gradient_checkpointing(type::String="speed")

Uses checkpointing scheme for gradients.

'speed': checkpoint all outputs of convolutions and matmuls. these ops are usually the most expensive, so checkpointing them maximizes the running speed (this is a good option if nonlinearities, concats, batchnorms, etc are taking up a lot of memory)

'memory': try to minimize the memory usage (currently using a very simple strategy that identifies a number of bottleneck tensors in the graph to checkpoint)

'collection': look for a tensorflow collection named 'checkpoints', which holds the tensors to checkpoint

source

ADCME.gradients — Method.

gradients(ys::PyObject, xs::PyObject; kwargs...)

Computes the gradients of ys w.r.t xs.

If ys is a scalar, gradients returns the gradients with the same shape as xs.

If ys is a vector, gradients returns the Jacobian $\frac{\partial y}{\partial x}$

!!!

The second usage is not suggested since `ADCME` adopts reverse mode automatic differentiation. 
Although in the case `ys` is a vector and `xs` is a scalar, `gradients` cleverly uses forward mode automatic differentiation,
it requires that the second order gradients are implemented for relevant operators. 

source

ADCME.hessian — Method.

hessian computes the hessian of a scalar function f with respect to vector inputs xs

source

ADCME.tensor — Method.

tensor(v::Array{T,2}; dtype=Float64, sparse=false) where T

source

ADCME.tensor — Method.

tensor(v::Array{T,2}; dtype=Float64, sparse=false) where T

Convert a generic array v to a tensor. For example,

v = [0.0 constant(1.0) 2.0
    constant(2.0) 0.0 1.0]
u = tensor(v)

u will be a $2\times 3$ tensor.

Note

This function is expensive. Use with caution.

source

Base.read — Method.

read(ta::PyObject, i::Union{PyObject,Integer})

Reads data from TensorArray at index i.

source

Base.write — Method.

write(ta::PyObject, i::Union{PyObject,Integer}, obj)

Writes data obj to TensorArray at index i.

source

Random Variables

ADCME.categorical — Method.

categorical(n::Union{PyObject, Integer}; kwargs...)

kwargs has a keyword argument logits, a 2-D Tensor with shape [batch_size, num_classes]. Each slice [i, :] represents the unnormalized log-probabilities for all classes.

source

ADCME.choice — Method.

choice(inputs::Union{PyObject, Array}, n_samples::Union{PyObject, Integer};replace::Bool=false)

Choose n_samples samples from inputs with/without replacement.

source

Sparse Matrix

ADCME.SparseTensor — Method.

SparseTensor(A::SparseMatrixCSC)

source

ADCME.SparseTensor — Method.

SparseTensor(I::Union{PyObject,Array{T,1}}, J::Union{PyObject,Array{T,1}}, V::Union{Array{Float64,1}, PyObject}, m::Union{S, PyObject, Nothing}=nothing, n::Union{S, PyObject, Nothing}=nothing) where {T<:Integer, S<:Integer}

Constructs a sparse tensor. Examples:

ii = [1;2;3;4]
jj = [1;2;3;4]
vv = [1.0;1.0;1.0;1.0]
s = SparseTensor(ii, jj, vv, 4, 4)
s = SparseTensor(sprand(10,10,0.3))

source

ADCME.SparseAssembler — Method.

accumulator, creater, initializer = SparseAssembler()

Returns 3 functions that can be used for assembling sparse matrices concurrently.

initializer must be called before the working session

accumulator accumulates column indices and values

creator accepts no input and outputs row indices, column indices and values for the sparse matrix

Example:

accumulator, creater, initializer = SparseAssembler()
initializer(5)
op1 = accumulator(1, [1;2;3], ones(3))
op2 = accumulator(1, [3], [1.])
op3 = accumulator(2, [1;3], ones(2))
run(sess, [op1,op2,op3])
ii,jj,vv = creater()
i,j,v = run(sess, [ii,jj,vv])
A = sparse(i,j,v,5,5)
@assert Array(A)≈[1.0  1.0  2.0  0.0  0.0
                1.0  0.0  1.0  0.0  0.0
                0.0  0.0  0.0  0.0  0.0
                0.0  0.0  0.0  0.0  0.0
                0.0  0.0  0.0  0.0  0.0]

source

ADCME.find — Method.

find(s::SparseTensor)

Returns the row, column and values for sparse tensor s.

source

ADCME.spdiag — Method.

spdiag(n::Int64)

Constructs a sparse identity matrix of size $n\times n$.

source

ADCME.spdiag — Method.

spdiag(o::PyObject)

Constructs a sparse diagonal matrix where the diagonal entries are o

source

ADCME.spzero — Function.

spzero(m::Int64, n::Union{Missing, Int64}=missing)

Constructs a empty sparse matrix of size $m\times n$. n=m if n is missing

source

Operations

ADCME.vector — Method.

vector(i::Union{Array{T}, PyObject, UnitRange, StepRange}, v::Union{Array{Float64},PyObject},s::Union{Int64,PyObject})

Returns a vector V with length s such that

V[i] = v

source

LinearAlgebra.svd — Method.

svd(o::PyObject, args...; kwargs...)

Returns a TFSVD structure which holds the following data structures

S::PyObject
U::PyObject
V::PyObject
Vt::PyObject

We have the equality $o = USV'$

source

IO

ADCME.Diary — Type.

Diary(suffix::Union{String, Nothing}=nothing)

Creates a diary at a temporary directory path. It returns a writer and the corresponding directory path

source

ADCME.activate — Function.

activate(sw::Diary, port::Int64=6006)

Running Diary at http://localhost:port.

source

ADCME.load — Function.

load(sess::PyObject, file::String, vars::Union{PyObject, Nothing, Array{PyObject}}=nothing, args...; kwargs...)

Loads the values of variables to the session sess from the file file. If vars is nothing, it loads values to all the trainable variables. See also save, load

source

ADCME.load — Method.

load(sw::Diary, dirp::String)

Loads Diary from dirp.

source

ADCME.pload — Method.

pload(file::String)

Loads a Python objection from file. See also psave

source

ADCME.psave — Method.

psave(o::PyObject, file::String)

Saves a Python objection o to file. See also pload

source

ADCME.save — Function.

save(sess::PyObject, file::String, vars::Union{PyObject, Nothing, Array{PyObject}}=nothing, args...; kwargs...)

Saves the values of vars in the session sess. The result is written into file as a dictionary. If vars is nothing, it saves all the trainable variables. See also save, load

source

ADCME.save — Method.

save(sw::Diary, dirp::String)

Saves Diary to dirp.

source

ADCME.scalar — Function.

scalar(o::PyObject, name::String)

Returns a scalar summary object.

source

Base.write — Method.

write(sw::Diary, step::Int64, cnt::Union{String, Array{String}})

Writes to Diary.

source

Optimization

ADCME.BFGS! — Function.

BFGS!(sess::PyObject, loss::PyObject, max_iter::Int64=15000; kwargs...)

BFGS! is a simplified interface for BFGS optimizer. See also ScipyOptimizerInterface.

example

a = Variable(1.0)
loss = (a - 10.0)^2
BFGS!(sess, loss)

source

ADCME.BFGS! — Function.

BFGS!(value_and_gradients_function::Function, initial_position::Union{PyObject, Array{Float64}}, max_iter::Int64=50, args...;kwargs...)

Applies the BFGS optimizer to value_and_gradients_function

source

ADCME.BFGS! — Method.

BFGS!(sess::PyObject, loss::PyObject, grads::Union{Array{T},Nothing,PyObject}, 
    vars::Union{Array{PyObject},PyObject}; kwargs...) where T<:Union{Nothing, PyObject}

Running BFGS algorithm $\min_{\texttt{vars}} \texttt{loss}(\texttt{vars})$ The gradients grads must be provided. Typically, grads[i] = gradients(loss, vars[i]). grads[i] can exist on different devices (GPU or CPU).

source

ADCME.CustomOptimizer — Method.

CustomOptimizer(opt::Function, name::String)

creates a custom optimizer with struct name name. For example, we can integrate Optim.jl with ADCME by constructing a new optimizer

CustomOptimizer("Con") do f, df, c, dc, x0, nineq, neq, x_L, x_U
    opt = Opt(:LD_MMA, length(x0))
    bd = zeros(length(x0)); bd[end-1:end] = [-Inf, 0.0]
    opt.lower_bounds = bd
    opt.xtol_rel = 1e-4
    opt.min_objective = (x,g)->(g[:]= df(x); return f(x)[1])
    inequality_constraint!(opt, (x,g)->( g[:]= dc(x);c(x)[1]), 1e-8)
    (minf,minx,ret) = NLopt.optimize(opt, x0)
    minx
end

Then we can create an optimizer with

opt = Con(loss, inequalities=[c1], equalities=[c2])

To trigger the optimization, use

opt.minimize(sess)

or

minimize(opt, sess)

Note thanks to the global variable scope of Julia, step_callback, optimizer_kwargs can actually be passed from Julia environment directly.

source

ADCME.NonlinearConstrainedProblem — Method.

NonlinearConstrainedProblem(f::Function, L::Function, θ::PyObject, u0::Union{PyObject, Array{Float64}}; options::Union{Dict{String, T}, Missing}=missing) where T<:Integer

Computes the gradients $\frac{\partial L}{\partial \theta}$

\[\begin{align} \min & \ L(u)\\ \mathrm{s.t.} & \ F(\theta, u) = 0 \end{align}\]

u0 is the initial guess for the numerical solution u, see newton_raphson.

Caveats: Assume r, A = f(θ, u) and θ are the unknown parameters, gradients(r, θ) must be defined (backprop works properly)

Returns: It returns a tuple (L: loss, C: constraints, and Graidents)

\[\left(L(u), u, \frac{\partial L}{\partial θ}\right)\]

source

ADCME.ScipyOptimizerInterface — Method.

ScipyOptimizerInterface(loss; method="L-BFGS-B", options=Dict("maxiter"=> 15000, "ftol"=>1e-12, "gtol"=>1e-12), kwargs...)

A simple interface for Scipy Optimizer. See also ScipyOptimizerMinimize and BFGS!.

source

ADCME.ScipyOptimizerMinimize — Method.

ScipyOptimizerMinimize(sess::PyObject, opt::PyObject; kwargs...)

Minimizes a scalar Tensor. Variables subject to optimization are updated in-place at the end of optimization.

Note that this method does not just return a minimization Op, unlike minimize; instead it actually performs minimization by executing commands to control a Session https://www.tensorflow.org/api_docs/python/tf/contrib/opt/ScipyOptimizerInterface. See also ScipyOptimizerInterface and BFGS!.

feed_dict: A feed dict to be passed to calls to session.run.

fetches: A list of Tensors to fetch and supply to loss_callback as positional arguments.

step_callback: A function to be called at each optimization step; arguments are the current values of all optimization variables flattened into a single vector.

loss_callback: A function to be called every time the loss and gradients are computed, with evaluated fetches supplied as positional arguments.

run_kwargs: kwargs to pass to session.run.

source

ADCME.newton_raphson — Method.

newton_raphson(f::Function, u::Union{Array,PyObject}, θ::Union{Missing,PyObject}; options::Union{Dict{String, T}, Missing}=missing)

Newton Raphson solver for solving a nonlinear equation. f has the signature f(u::PyObject, θ::Union{Missing,PyObject})->(r::PyObject, A::Union{PyObject,SparseTensor}) where r is the residual and A is the Jacobian matrix. θ are external parameters. u0 is the initial guess for u options:

"max_iter": maximum number of iterations (default=100)

"verbose": whether details are printed (default=false)

"rtol": relative tolerance for termination (default=1e-12)

"tol": absolute tolerance for termination (default=1e-12)

source

Neural Networks

ADCME.ae — Function.

ae(x::PyObject, output_dims::Array{Int64}, scope::String = "default")

Creates a neural network with intermediate numbers of neurons output_dims.

source

ADCME.ae — Method.

ae(x::Union{Array{Float64}, PyObject}, output_dims::Array{Int64}, θ::Union{Array{Float64}, PyObject})

Creates a neural network with intermediate numbers of neurons output_dims. The weights are given by θ

source

ADCME.ae_init — Method.

ae_init(sess::PyObject, x::Union{Array{Float64}, PyObject}, output_dims::Array{Int64})

Return the initial weights and bias values by TensorFlow as a vector.

source

ADCME.ae_to_code — Method.

ae_to_code(file::String, scope::String)

Return the code string from the feed-forward neural network data in file.

source

ADCME.bn — Method.

bn(args...;center = true, scale=true, kwargs...)

example

bn(inputs, name="batch_norm", is_training=true)

Note

bn should be used with control_dependency

update_ops = get_collection(UPDATE_OPS)
control_dependencies(update_ops) do 
    global train_step = AdamOptimizer().minimize(loss)
end 

source

ADCME.num_ae — Method.

num_ae(output_dims::Array{Int64})

Estimates the number of weights for the neural network.

source

Generative Neural Nets

ADCME.GAN — Type.

GAN(dat::PyObject, 
    generator::Function, 
    gan::GAN,
    loss::Union{String, Function, Missing}=missing; 
    latent_dim::Union{Missing, Int64}=missing, 
    batch_size::Union{Missing, Int64}=missing)

Creates a GAN instance.

dat $\in \mathbb{R}^{n\times d}$ is the training data for the GAN, where $n$ is the number of training data, and $d$ is the dimension per training data.

generator$:\mathbb{R}^{d'} \rightarrow \mathbb{R}^d$ is the generator function, $d'$ is the hidden dimension.

discriminator$:\mathbb{R}^{d} \rightarrow \mathbb{R}$ is the discriminator function.

loss is the loss function. See klgan, rklgan, wgan, lsgan for examples.

latent_dim (default=$d$) is the latent dimension.

batch_size (default=32) is the batch size in training.

source

ADCME.jsgan — Method.

jsgan(gan::GAN)

Computes the vanilla GAN loss function.

source

ADCME.klgan — Method.

klgan(gan::GAN)

Computes the KL-divergence GAN loss function.

source

ADCME.lsgan — Method.

lsgan(gan::GAN)

Computes the least square GAN loss function.

source

ADCME.predict — Method.

predict(gan::GAN, input::Union{PyObject, Array})

Predicts the GAN gan output given input input.

source

ADCME.rklgan — Method.

rklgan(gan::GAN)

Computes the reverse KL-divergence GAN loss function.

source

ADCME.sample — Method.

sample(gan::GAN, n::Int64)

Samples n instances from gan.

source

ADCME.wgan — Method.

wgan(gan::GAN)

Computes the Wasserstein GAN loss function.

source

ADCME.build! — Method.

build!(gan::GAN)

Builds the GAN instances. This function returns gan for convenience.

source

Tools

ADCME.compile_op — Method.

compile_op(oplibpath::String; check::Bool=false)

Compile the library operator by force.

source

ADCME.customop — Method.

customop()

Create a new custom operator.

example

julia> customop() # create an editable `customop.txt` file
[ Info: Edit custom_op.txt for custom operators
julia> customop() # after editing `customop.txt`, call it again to generate interface files.

source

ADCME.load_op — Method.

load_op(oplibpath::String, opname::String)

Loads the operator opname from library oplibpath.

source

ADCME.load_op_and_grad — Method.

load_op_and_grad(oplibpath::String, opname::String; multiple::Bool=false)

Loads the operator opname from library oplibpath; gradients are also imported. If multiple is true, the operator is assumed to have multiple outputs.

source

ADCME.xavier_init — Function.

xavier_init(size, dtype=Float64)

Returns a matrix of size size and its values are from Xavier initialization.

source

ADCME.compile — Method.

compile(s::String)

Compiles the library s by force.

source

ADCME.load_system_op — Function.

load_system_op(s::String, oplib::String, grad::Bool=true)

Loads custom operator from CustomOps directory (shipped with ADCME instead of TensorFlow) For example

s = "SparseOperator"
oplib = "libSO"
grad = true

this will direct Julia to find library CustomOps/SparseOperator/libSO.dylib on MACOSX

source

Datasets