Additional Tools

Additional Tools

There are many handy tools implemented in ADCME for analysis, benchmarking, input/output, etc.

Benchmarking

The functions tic and toc can be used for recording the runtime between two operations. tic starts a timer for performance measurement while toc marks the termination of the measurement. Both functions are bound with one operations. For example, we can benchmark the runtime for svd

A = constant(rand(10,20))
A = tic(A)
r = svd(A)
B = r.U*diagm(r.S)*r.Vt 
B, t = toc(B)
run(sess, B)
run(sess, t)

ADCME.tic — Function.

tic(o::PyObject, i::Union{PyObject, Integer}=0)

Construts a TensorFlow timer with index i. The start time record is right before o is executed.

source

ADCME.toc — Function.

toc(o::PyObject, i::Union{PyObject, Integer}=0)

Returns the elapsed time from last tic call with index i (default=0). The terminal time record is right before o is executed.

source

Save and Load Python Object

ADCME.psave — Function.

psave(o::PyObject, file::String)

Saves a Python objection o to file. See also pload

source

ADCME.pload — Function.

pload(file::String)

Loads a Python objection from file. See also psave

source

Save and Load TensorFlow Session

ADCME.load — Function.

load(sess::PyObject, file::String, vars::Union{PyObject, Nothing, Array{PyObject}}=nothing, args...; kwargs...)

Loads the values of variables to the session sess from the file file. If vars is nothing, it loads values to all the trainable variables. See also save, load

source

load(sw::Diary, dirp::String)

Loads Diary from dirp.

source

ADCME.save — Function.

save(sess::PyObject, file::String, vars::Union{PyObject, Nothing, Array{PyObject}}=nothing, args...; kwargs...)

Saves the values of vars in the session sess. The result is written into file as a dictionary. If vars is nothing, it saves all the trainable variables. See also save, load

source

save(sw::Diary, dirp::String)

Saves Diary to dirp.

source

Save and Load Diary

We can use TensorBoard to track a scalar value easily

d = Diary("test")
p = placeholder(1.0, dtype=Float64)
b = constant(1.0)+p
s = scalar(b, "variable")
for i = 1:100
    write(d, i, run(sess, s, Dict(p=>Float64(i))))
end
activate(d)

ADCME.Diary — Type.

Diary(suffix::Union{String, Nothing}=nothing)

Creates a diary at a temporary directory path. It returns a writer and the corresponding directory path

source

ADCME.scalar — Function.

scalar(o::PyObject, name::String)

Returns a scalar summary object.

source

ADCME.activate — Function.

activate(sw::Diary, port::Int64=6006)

Running Diary at http://localhost:port.

source

Missing docstring.

Missing docstring for load. Check Documenter's build log for details.

Missing docstring.

Missing docstring for save. Check Documenter's build log for details.

Base.write — Function.

write(sw::Diary, step::Int64, cnt::Union{String, Array{String}})

Writes to Diary.

source