API Reference

Core Functions

ADCME.add_collection — Method

add_collection(name::String, v::PyObject)

Adds v to the collection with name name. If name does not exist, a new one is created.

ADCME.add_collection — Method

add_collection(name::String, vs::PyObject...)

Adds operators vs to the collection with name name. If name does not exist, a new one is created.

ADCME.control_dependencies — Method

control_dependencies(f, ops::Union{Array{PyObject}, PyObject})

Executes all operations in ops before any operations created inside the block.

op1 = tf.print("print op1")
op3 = tf.print("print op3")
control_dependencies(op1) do
    global op2 = tf.print("print op2")
end
run(sess, [op2,op3])

In this example, op1 must be executed before op2. But there is no guarantee when op3 will be executed. There are several possible outputs of the program such as

print op3
print op1
print op2

or

print op1
print op3
print op2

ADCME.get_collection — Function

get_collection(name::Union{String, Missing})

Returns the collection with name name. If name is missing, returns all the trainable variables.

ADCME.has_gpu — Method

has_gpu()

Checks if GPU is available.

Note

ADCME will use GPU automatically if GPU is available. To disable GPU, set the environment variable ENV["CUDA_VISIBLE_DEVICES"]="" before importing ADCME

ADCME.if_else — Method

if_else(condition::Union{PyObject,Array,Bool}, fn1, fn2, args...;kwargs...)

If condition is a scalar boolean, it outputs fn1 or fn2 (a function with no input argument or a tensor) based on whether condition is true or false.

If condition is a boolean array, if returns condition .* fn1 + (1 - condition) .* fn2

ADCME.independent — Method

independent(o::PyObject, args...; kwargs...)

Returns o but when computing the gradients, the top gradients will not be back-propagated into dependent variables of o.

ADCME.reset_default_graph — Method

reset_default_graph()

Resets the graph by removing all the operators.

ADCME.tensor — Method

tensor(s::String)

Returns the tensor with name s. See tensorname.

ADCME.tensorname — Method

tensorname(o::PyObject)

Returns the name of the tensor. See tensor.

ADCME.while_loop — Method

while_loop(condition::Union{PyObject,Function}, body::Function, loop_vars::Union{PyObject, Array{Any}, Array{PyObject}};
    parallel_iterations::Int64=10, kwargs...)

Loops over loop_vars while condition is true. This operator only creates one extra node to mark the loops in the computational graph.

Example

The following script computes

\[\sum_{i=1}^10 i\]

function condition(i, ta)
    i <= 10
end
function body(i, ta)
    u = read(ta, i-1)
    ta = write(ta, i, u+1)
    i+1, ta
end
ta = TensorArray(10)
ta = write(ta, 1, constant(1.0))
i = constant(2, dtype=Int32)
_, out = while_loop(condition, body, [i, ta])
summation = stack(out)[10]

ADCME.run_profile — Method

run_profile(args...;kwargs...)

Runs the session with tracing information.

ADCME.save_profile — Function

save_profile(filename::String="default_timeline.json")

Save the timeline information to file filename.

Open Chrome and navigate to chrome://tracing

Load the timeline file

Base.bind — Method

bind(op::PyObject, ops...)

Adding operations ops to the dependencies of op. The function is useful when we want to execute ops but ops is not in the dependency of the final output. For example, if we want to print i each time i is evaluated

i = constant(1.0)
op = tf.print(i)
i = bind(i, op)

Variables

ADCME.TensorArray — Function

TensorArray(size_::Int64=0, args...;kwargs...)

Constructs a tensor array for while_loop.

ADCME.Variable — Method

Variable(initial_value;kwargs...)

Constructs a ref tensor from value.

ADCME.cell — Method

cell(arr::Array, args...;kwargs...)

Construct a cell tensor.

Example

julia> r = cell([[1.],[2.,3.]])
julia> run(sess, r[1])
1-element Array{Float32,1}:
 1.0
julia> run(sess, r[2])
2-element Array{Float32,1}:
 2.0
 3.0

ADCME.constant — Method

constant(value; kwargs...)

Constructs a non-trainable tensor from value.

ADCME.convert_to_tensor — Method

convert_to_tensor(o::Union{PyObject, Number, Array{T}, Missing, Nothing}; dtype::Union{Type, Missing}=missing) where T<:Number
convert_to_tensor(os::Array, dtypes::Array)

Converts the input o to tensor. If o is already a tensor and dtype (if provided) is the same as that of o, the operator does nothing. Otherwise, convert_to_tensor converts the numerical array to a constant tensor or casts the data type. convert_to_tensor also accepts multiple tensors.

Example

convert_to_tensor([1.0, constant(rand(2)), rand(10)], [Float32, Float64, Float32])

ADCME.gradient_checkpointing — Function

gradient_checkpointing(type::String="speed")

Uses checkpointing scheme for gradients.

'speed': checkpoint all outputs of convolutions and matmuls. these ops are usually the most expensive, so checkpointing them maximizes the running speed (this is a good option if nonlinearities, concats, batchnorms, etc are taking up a lot of memory)

'memory': try to minimize the memory usage (currently using a very simple strategy that identifies a number of bottleneck tensors in the graph to checkpoint)

'collection': look for a tensorflow collection named 'checkpoints', which holds the tensors to checkpoint

ADCME.gradients — Method

gradients(ys::PyObject, xs::PyObject; kwargs...)

Computes the gradients of ys w.r.t xs.

If ys is a scalar, gradients returns the gradients with the same shape as xs.

If ys is a vector, gradients returns the Jacobian $\frac{\partial y}{\partial x}$

Note

The second usage is not suggested since ADCME adopts reverse mode automatic differentiation. Although in the case ys is a vector and xs is a scalar, gradients cleverly uses forward mode automatic differentiation, it requires that the second order gradients are implemented for relevant operators.

ADCME.hessian — Method

hessian computes the hessian of a scalar function f with respect to vector inputs xs

ADCME.tensor — Method

tensor(v::Array{T,2}; dtype=Float64, sparse=false) where T

ADCME.tensor — Method

tensor(v::Array{T,2}; dtype=Float64, sparse=false) where T

Convert a generic array v to a tensor. For example,

v = [0.0 constant(1.0) 2.0
    constant(2.0) 0.0 1.0]
u = tensor(v)

u will be a $2\times 3$ tensor.

Note

This function is expensive. Use with caution.

Base.read — Method

read(ta::PyObject, i::Union{PyObject,Integer})

Reads data from TensorArray at index i.

Base.write — Method

write(ta::PyObject, i::Union{PyObject,Integer}, obj)

Writes data obj to TensorArray at index i.

Random Variables

ADCME.categorical — Method

categorical(n::Union{PyObject, Integer}; kwargs...)

kwargs has a keyword argument logits, a 2-D Tensor with shape [batch_size, num_classes]. Each slice [i, :] represents the unnormalized log-probabilities for all classes.

ADCME.choice — Method

choice(inputs::Union{PyObject, Array}, n_samples::Union{PyObject, Integer};replace::Bool=false)

Choose n_samples samples from inputs with/without replacement.

Sparse Matrix

ADCME.SparseTensor — Method

SparseTensor(A::SparseMatrixCSC)
SparseTensor(A::Array{Float64, 2})

Creates a SparseTensor from numerical arrays.

ADCME.SparseTensor — Method

SparseTensor(I::Union{PyObject,Array{T,1}}, J::Union{PyObject,Array{T,1}}, V::Union{Array{Float64,1}, PyObject}, m::Union{S, PyObject, Nothing}=nothing, n::Union{S, PyObject, Nothing}=nothing) where {T<:Integer, S<:Integer}

Constructs a sparse tensor. Examples:

ii = [1;2;3;4]
jj = [1;2;3;4]
vv = [1.0;1.0;1.0;1.0]
s = SparseTensor(ii, jj, vv, 4, 4)
s = SparseTensor(sprand(10,10,0.3))

ADCME.SparseAssembler — Function

SparseAssembler(handle::Union{PyObject, <:Integer}, n::Union{PyObject, <:Integer}, tol::Union{PyObject, <:Real}=0.0)

Creates a SparseAssembler for accumulating row, col, val for sparse matrices.

handle: an integer handle for creating a sparse matrix. If the handle already exists, SparseAssembler return the existing sparse matrix handle. If you are creating different sparse matrices, the handles should be different.

n: Number of rows of the sparse matrix.

tol (optional): Tolerance. SparseAssembler will treats any values less than tol as zero.

Example 1

handle = SparseAssembler(100, 5, 1e-8)
op1 = accumulate(handle, 1, [1;2;3], [1.0;2.0;3.0])
op2 = accumulate(handle, 2, [1;2;3], [1.0;2.0;3.0])
J = assemble(5, 5, [op1;op2])

J will be a SparseTensor object.

Example 2

handle = SparseAssembler(0, 5)
op1 = accumulate(handle, 1, [1;2;3], ones(3))
op2 = accumulate(handle, 1, [3], [1.])
op3 = accumulate(handle, 2, [1;3], ones(2))
J = assemble(5, 5, [op1;op2;op3]) # op1, op2, op3 are parallel
Array(run(sess, J))≈[1.0  1.0  2.0  0.0  0.0
                1.0  0.0  1.0  0.0  0.0
                0.0  0.0  0.0  0.0  0.0
                0.0  0.0  0.0  0.0  0.0
                0.0  0.0  0.0  0.0  0.0]

ADCME.assemble — Method

assemble(m::Union{PyObject, <:Integer}, n::Union{PyObject, <:Integer}, ops::PyObject)

Assembles the sparse matrix from the ops created by accumulate. ops is either a single output from accumulate, or concated from several ops

op1 = accumulate(handle, 1, [1;2;3], [1.0;2.0;3.0])
op2 = accumulate(handle, 2, [1;2;3], [1.0;2.0;3.0])
op = [op1;op2] # equivalent to `vcat([op1, op2]...)`

m and n are rows and columns of the sparse matrix.

See SparseAssembler for an example.

ADCME.find — Method

find(s::SparseTensor)

Returns the row, column and values for sparse tensor s.

ADCME.scatter_add — Method

scatter_update(A::Union{SparseTensor, SparseMatrixCSC{Float64,Int64}},
i1::Union{Integer, Colon, UnitRange{T}, PyObject,Array{S,1}},
i2::Union{Integer, Colon, UnitRange{T}, PyObject,Array{T,1}},
B::Union{SparseTensor, SparseMatrixCSC{Float64,Int64}})  where {S<:Real,T<:Real}

Adds B to a subblock of a sparse matrix A. Equivalently,

A[i1, i2] += B

ADCME.scatter_update — Method

scatter_update(A::Union{SparseTensor, SparseMatrixCSC{Float64,Int64}},
i1::Union{Integer, Colon, UnitRange{T}, PyObject,Array{S,1}},
i2::Union{Integer, Colon, UnitRange{T}, PyObject,Array{T,1}},
B::Union{SparseTensor, SparseMatrixCSC{Float64,Int64}})  where {S<:Real,T<:Real}

Updates a subblock of a sparse matrix by B. Equivalently,

A[i1, i2] = B

ADCME.spdiag — Method

spdiag(n::Int64)

Constructs a sparse identity matrix of size $n\times n$.

ADCME.spdiag — Method

spdiag(m::Integer, pair::Pair...)

Constructs a square $m\times m$ SparseTensor from pairs of the form

offset => array 

ADCME.spdiag — Method

spdiag(o::PyObject)

Constructs a sparse diagonal matrix where the diagonal entries are o

ADCME.spzero — Function

spzero(m::Int64, n::Union{Missing, Int64}=missing)

Constructs a empty sparse matrix of size $m\times n$. n=m if n is missing

Base.accumulate — Method

accumulate(handle::PyObject, row::Union{PyObject, <:Integer}, cols::Union{PyObject, Array{<:Integer}}, vals::Union{PyObject, Array{<:Real}})

Accumulates row-th row. It adds the value to the sparse matrix

for k = 1:length(cols)
    A[row, cols[k]] += vals[k]
end

handle is the handle created by SparseAssembler.

See SparseAssembler for an example.

Note

The function accumulate returns a op::PyObject. Only when op is executed, the nonzero values are populated into the sparse matrix.

Base.:\ — Function

\(s::SparseTensor, o::PyObject, method::String="SparseLU")

Solves the linear equation $s x = o$

Method

For square matrices s, one of the following methods is available

SparseLU

SparseQR

SimplicialLDLT

SimplicialLLT

Operations

ADCME.argsort — Method

argsort(o::PyObject; 
stable::Bool = false, rev::Bool=false, dims::Integer=-1, name::Union{Nothing,String}=nothing)

Returns the indices of a tensor that give its sorted order along an axis.

ADCME.batch_matmul — Method

batch_matmul(o1::PyObject, o2::PyObject)

Computes o1[i,:,:] * o2[i, :] or o1[i,:,:] * o2[i, :, :] for each index i.

ADCME.clip — Method

clip(o::Union{Array{Any}, Array{PyObject}}, vmin, vmax, args...;kwargs...)

Clips the values of o to the range [vmin, vmax]

Example

a = constant(3.0)
a = clip(a, 1.0, 2.0)
b = constant(rand(3))
b = clip(b, 0.5, 1.0)

ADCME.cvec — Method

rvec(o::PyObject; kwargs...)

Vectorizes the tensor o to a column vector, assuming column major.

ADCME.pmap — Method

pmap(fn::Function, o::Union{Array{PyObject}, PyObject})

Parallel for loop. There should be no data dependency between different iterations.

Example

x = constant(ones(10))
y1 = pmap(x->2.0*x, x)
y2 = pmap(x->x[1]+x[2], [x,x])
y3 = pmap(1:10, x) do z
    i = z[1]
    xi = z[2]
    xi + cast(Float64, i)
end
run(sess, y1)
run(sess, y2)
run(sess, y3)

ADCME.rvec — Method

rvec(o::PyObject; kwargs...)

Vectorizes the tensor o to a row vector, assuming row major.

ADCME.set_shape — Method

set_shape(o::PyObject, s::Union{Array{<:Integer}, Tuple{Vararg{<:Integer, N}}}) where N
set_shape(o::PyObject, s::Integer...)

Sets the shape of o to s. s must be the actual shape of o. This function is used to convert a tensor with unknown dimensions to a tensor with concrete dimensions.

Example

a = placeholder(Float64, shape=[nothing, 10])
b = set_shape(a, 3, 10)
run(sess, b, a=>rand(3,10)) # OK 
run(sess, b, a=>rand(5,10)) # Error
run(sess, b, a=>rand(10,3)) # Error

ADCME.topk — Function

topk(o::PyObject, k::Union{PyObject,Integer}=1;
    sorted::Bool=true, name::Union{Nothing,String}=nothing)

Finds values and indices of the k largest entries for the last dimension. If sorted=true the resulting k elements will be sorted by the values in descending order.

ADCME.vector — Method

vector(i::Union{Array{T}, PyObject, UnitRange, StepRange}, v::Union{Array{Float64},PyObject},s::Union{Int64,PyObject})

Returns a vector V with length s such that

V[i] = v

Base.vec — Method

vec(o::PyObject;kwargs...)

Vectorizes the tensor o assuming column major.

LinearAlgebra.svd — Method

svd(o::PyObject, args...; kwargs...)

Returns a TFSVD structure which holds the following data structures

S::PyObject
U::PyObject
V::PyObject
Vt::PyObject

We have the equality $o = USV'$

Example

A = rand(10,20)
r = svd(constant(A))
A2 = r.U*diagm(r.S)*r.Vt # The value of `A2` should be equal to `A`

Base.map — Method

map(fn::Function, o::Union{Array{PyObject},PyObject};
kwargs...)

Applies fn to each element of o.

o∈Array{PyObject} : returns [fn(x) for x in o]

o∈PyObject : splits o according to the first dimension and then applies fn.

Example

a = constant(rand(10,5))
b = map(x->sum(x), a) # equivalent to `sum(a, dims=2)`

Note

If fn is a multivariate function, we need to specify the output type using dtype keyword. For example,

a = constant(ones(10))
b = constant(ones(10))
fn = x->x[1]+x[2]
c = map(fn, [a, b], dtype=Float64)

Base.reshape — Method

reshape(o::PyObject, s::Union{Array{<:Integer}, Tuple{Vararg{<:Integer, N}}}) where N 
reshape(o::PyObject, s::Integer; kwargs...)
reshape(o::PyObject, m::Integer, n::Integer; kwargs...)
reshape(o::PyObject, ::Colon, n::Integer)
reshape(o::PyObject, n::Integer, ::Colon)

Reshapes the tensor. reshape is compatible with

Base.sort — Method

Base.:sort(o::PyObject; 
rev::Bool=false, dims::Integer=-1, name::Union{Nothing,String}=nothing)

Sort a multidimensional array o along the given dimension.

rev: true for DESCENDING and false (default) for ASCENDING

dims: -1 for last dimension.

IO

ADCME.Diary — Type

Diary(suffix::Union{String, Nothing}=nothing)

Creates a diary at a temporary directory path. It returns a writer and the corresponding directory path

ADCME.activate — Function

activate(sw::Diary, port::Int64=6006)

Running Diary at http://localhost:port.

ADCME.load — Function

load(sess::PyObject, file::String, vars::Union{PyObject, Nothing, Array{PyObject}}=nothing, args...; kwargs...)

Loads the values of variables to the session sess from the file file. If vars is nothing, it loads values to all the trainable variables. See also save, load

ADCME.load — Method

load(sw::Diary, dirp::String)

Loads Diary from dirp.

ADCME.pload — Method

pload(file::String)

Loads a Python objection from file. See also psave

ADCME.psave — Method

psave(o::PyObject, file::String)

Saves a Python objection o to file. See also pload

ADCME.save — Function

save(sess::PyObject, file::String, vars::Union{PyObject, Nothing, Array{PyObject}}=nothing, args...; kwargs...)

Saves the values of vars in the session sess. The result is written into file as a dictionary. If vars is nothing, it saves all the trainable variables. See also save, load

ADCME.save — Method

save(sw::Diary, dirp::String)

Saves Diary to dirp.

ADCME.scalar — Function

scalar(o::PyObject, name::String)

Returns a scalar summary object.

Base.write — Method

write(sw::Diary, step::Int64, cnt::Union{String, Array{String}})

Writes to Diary.

Optimization

ADCME.BFGS! — Function

BFGS!(value_and_gradients_function::Function, initial_position::Union{PyObject, Array{Float64}}, max_iter::Int64=50, args...;kwargs...)

Applies the BFGS optimizer to value_and_gradients_function

ADCME.BFGS! — Function

BFGS!(sess::PyObject, loss::PyObject, max_iter::Int64=15000; 
vars::Array{PyObject}=PyObject[], callback::Union{Function, Nothing}=nothing, kwargs...)

BFGS! is a simplified interface for BFGS optimizer. See also ScipyOptimizerInterface. callback is a callback function with signature

callback(vs::Array, iter::Int64, loss::Float64)

vars is an array consisting of tensors and its values will be the input to vs.

example

a = Variable(1.0)
loss = (a - 10.0)^2
BFGS!(sess, loss)

ADCME.BFGS! — Method

BFGS!(sess::PyObject, loss::PyObject, grads::Union{Array{T},Nothing,PyObject}, 
    vars::Union{Array{PyObject},PyObject}; kwargs...) where T<:Union{Nothing, PyObject}

Running BFGS algorithm $\min_{\texttt{vars}} \texttt{loss}(\texttt{vars})$ The gradients grads must be provided. Typically, grads[i] = gradients(loss, vars[i]). grads[i] can exist on different devices (GPU or CPU).

ADCME.CustomOptimizer — Method

CustomOptimizer(opt::Function, name::String)

creates a custom optimizer with struct name name. For example, we can integrate Optim.jl with ADCME by constructing a new optimizer

CustomOptimizer("Con") do f, df, c, dc, x0, nineq, neq, x_L, x_U
    opt = Opt(:LD_MMA, length(x0))
    bd = zeros(length(x0)); bd[end-1:end] = [-Inf, 0.0]
    opt.lower_bounds = bd
    opt.xtol_rel = 1e-4
    opt.min_objective = (x,g)->(g[:]= df(x); return f(x)[1])
    inequality_constraint!(opt, (x,g)->( g[:]= dc(x);c(x)[1]), 1e-8)
    (minf,minx,ret) = NLopt.optimize(opt, x0)
    minx
end

Then we can create an optimizer with

opt = Con(loss, inequalities=[c1], equalities=[c2])

To trigger the optimization, use

opt.minimize(sess)

or

minimize(opt, sess)

Note thanks to the global variable scope of Julia, step_callback, optimizer_kwargs can actually be passed from Julia environment directly.

ADCME.NonlinearConstrainedProblem — Method

NonlinearConstrainedProblem(f::Function, L::Function, θ::PyObject, u0::Union{PyObject, Array{Float64}}; options::Union{Dict{String, T}, Missing}=missing) where T<:Integer

Computes the gradients $\frac{\partial L}{\partial \theta}$

\[\min \ L(u) \quad \mathrm{s.t.} \ F(\theta, u) = 0\]

u0 is the initial guess for the numerical solution u, see newton_raphson.

Caveats: Assume r, A = f(θ, u) and θ are the unknown parameters, gradients(r, θ) must be defined (backprop works properly)

Returns: It returns a tuple (L: loss, C: constraints, and Graidents)

\[\left(L(u), u, \frac{\partial L}{\partial θ}\right)\]

ADCME.ScipyOptimizerInterface — Method

ScipyOptimizerInterface(loss; method="L-BFGS-B", options=Dict("maxiter"=> 15000, "ftol"=>1e-12, "gtol"=>1e-12), kwargs...)

A simple interface for Scipy Optimizer. See also ScipyOptimizerMinimize and BFGS!.

ADCME.ScipyOptimizerMinimize — Method

ScipyOptimizerMinimize(sess::PyObject, opt::PyObject; kwargs...)

Minimizes a scalar Tensor. Variables subject to optimization are updated in-place at the end of optimization.

Note that this method does not just return a minimization Op, unlike minimize; instead it actually performs minimization by executing commands to control a Session https://www.tensorflow.org/api_docs/python/tf/contrib/opt/ScipyOptimizerInterface. See also ScipyOptimizerInterface and BFGS!.

feed_dict: A feed dict to be passed to calls to session.run.

fetches: A list of Tensors to fetch and supply to loss_callback as positional arguments.

step_callback: A function to be called at each optimization step; arguments are the current values of all optimization variables flattened into a single vector.

loss_callback: A function to be called every time the loss and gradients are computed, with evaluated fetches supplied as positional arguments.

run_kwargs: kwargs to pass to session.run.

ADCME.newton_raphson — Method

newton_raphson(f::Function, u::Union{Array,PyObject}, θ::Union{Missing,PyObject}; options::Union{Dict{String, T}, Missing}=missing)

Newton Raphson solver for solving a nonlinear equation. f has the signature

f(θ::Union{Missing,PyObject}, u::PyObject)->(r::PyObject, A::Union{PyObject,SparseTensor}) (if linesearch is off)

f(θ::Union{Missing,PyObject}, u::PyObject)->(fval::PyObject, r::PyObject, A::Union{PyObject,SparseTensor}) (if linesearch is on)

where r is the residual and A is the Jacobian matrix; in the case where linesearch is on, the function value fval must also be supplied. θ are external parameters. u0 is the initial guess for u options:

max_iter: maximum number of iterations (default=100)

verbose: whether details are printed (default=false)

rtol: relative tolerance for termination (default=1e-12)

tol: absolute tolerance for termination (default=1e-12)

LM: a float number, Levenberg-Marquardt modification $x^{k+1} = x^k - (J^k + \mu^k)^{-1}g^k$ (default=0.0)

linesearch: whether linesearch is used (default=false)

Currently, the backtracing algorithm is implemented. The parameters for linesearch are also supplied via options

ls_c1: stop criterion, $f(x^k) < f(0) + \alpha c_1 f'(0)$

ls_ρ_hi: the new step size $\alpha_1\leq \rho_{hi}\alpha_0$

ls_ρ_lo: the new step size $\alpha_1\geq \rho_{lo}\alpha_0$

ls_iterations: maximum number of iterations for linesearch

ls_maxstep: maximum allowable steps

ls_αinitial: initial guess for the step size $\alpha$

Neural Networks

ADCME.ae — Function

ae(x::PyObject, output_dims::Array{Int64}, scope::String = "default";
    activation::Union{Function,String} = "tanh")

Creates a neural network with intermediate numbers of neurons output_dims.

ADCME.ae — Method

ae(x::Union{Array{Float64}, PyObject}, output_dims::Array{Int64}, θ::Union{Array{Float64}, PyObject};
activation::Union{Function,String} = "tanh")

Creates a neural network with intermediate numbers of neurons output_dims. The weights are given by θ

Example 1: Explicitly construct weights and biases

x = constant(rand(10,2))
n = ae_num([2,20,20,20,2])
θ = Variable(randn(n)*0.001)
y = ae(x, [20,20,20,2], θ)

Example 2: Implicitly construct weights and biases

θ = ae_init([10,20,20,20,2]) 
x = constant(rand(10,10))
y = ae(x, [20,20,20,2], θ)

See also ae_num, ae_init.

ADCME.ae_init — Method

ae_init(output_dims::Array{Int64}; T::Type=Float64, method::String="xavier")

Return the initial weights and bias values by TensorFlow as a vector. Three types of random initializers are provided

xavier (default). It is useful for tanh fully connected neural network.

\[W^l_i \sim \sqrt{\frac{1}{n_{l-1}}}\]

xavier_avg. A variant of xavier

\[W^l_i \sim \sqrt{\frac{2}{n_l + n_{l-1}}}\]

he. This is the activation aware initialization of weights and helps mitigate the problem

of vanishing/exploding gradients.

\[W^l_i \sim \sqrt{\frac{2}{n_{l-1}}}\]

ADCME.ae_num — Method

ae_num(output_dims::Array{Int64})

Estimates the number of weights and biases for the neural network. Note the first dimension should be the feature dimension (this is different from ae since in ae the feature dimension can be inferred), and the last dimension should be the output dimension.

ADCME.ae_to_code — Method

ae_to_code(file::String, scope::String; activation::String = "tanh")

Return the code string from the feed-forward neural network data in file. Usually we can immediately evaluate the code string into Julia session by

eval(Meta.parse(s))

If activation is not specified, tanh is the default.

ADCME.bn — Method

bn(args...;center = true, scale=true, kwargs...)

bn accepts a keyword parameter is_training.

Example

bn(inputs, name="batch_norm", is_training=true)

Note

bn should be used with control_dependency

update_ops = get_collection(UPDATE_OPS)
control_dependencies(update_ops) do 
    global train_step = AdamOptimizer().minimize(loss)
end 

Generative Neural Nets

ADCME.GAN — Type

GAN(dat::PyObject, 
    generator::Function, 
    discriminator::Function,
    loss::Union{String, Function, Missing}=missing; 
    latent_dim::Union{Missing, Int64}=missing, 
    batch_size::Union{Missing, Int64}=missing)

Creates a GAN instance.

dat $\in \mathbb{R}^{n\times d}$ is the training data for the GAN, where $n$ is the number of training data, and $d$ is the dimension per training data.

generator$:\mathbb{R}^{d'} \rightarrow \mathbb{R}^d$ is the generator function, $d'$ is the hidden dimension.

discriminator$:\mathbb{R}^{d} \rightarrow \mathbb{R}$ is the discriminator function.

loss is the loss function. See klgan, rklgan, wgan, lsgan for examples.

latent_dim (default=$d$, the same as output dimension) is the latent dimension.

batch_size (default=32) is the batch size in training.

ADCME.jsgan — Method

jsgan(gan::GAN)

Computes the vanilla GAN loss function.

ADCME.klgan — Method

klgan(gan::GAN)

Computes the KL-divergence GAN loss function.

ADCME.lsgan — Method

lsgan(gan::GAN)

Computes the least square GAN loss function.

ADCME.predict — Method

predict(gan::GAN, input::Union{PyObject, Array})

Predicts the GAN gan output given input input.

ADCME.rklgan — Method

rklgan(gan::GAN)

Computes the reverse KL-divergence GAN loss function.

ADCME.sample — Method

sample(gan::GAN, n::Int64)

Samples n instances from gan.

ADCME.wgan — Method

wgan(gan::GAN)

Computes the Wasserstein GAN loss function.

ADCME.build! — Method

build!(gan::GAN)

Builds the GAN instances. This function returns gan for convenience.

Tools

ADCME.compile_op — Method

compile_op(oplibpath::String; check::Bool=false)

Compile the library operator by force.

ADCME.customop — Function

customop(simple::Bool=false)

Create a new custom operator. If simple=true, the custom operator only supports CPU and does not have gradients.

Example

julia> customop() # create an editable `customop.txt` file
[ Info: Edit custom_op.txt for custom operators
julia> customop() # after editing `customop.txt`, call it again to generate interface files.

ADCME.install — Method

install(s::String; force::Bool = false)

Install a custom operator via URL. s can be

A URL. ADCME will download the directory through git

A string. ADCME will search for the associated package on https://github.com/ADCMEMarket

ADCME.load_op — Method

load_op(oplibpath::String, opname::String)

Loads the operator opname from library oplibpath.

ADCME.load_op_and_grad — Method

load_op_and_grad(oplibpath::String, opname::String; multiple::Bool=false)

Loads the operator opname from library oplibpath; gradients are also imported. If multiple is true, the operator is assumed to have multiple outputs.

ADCME.load_system_op — Function

load_system_op(s::String, oplib::String, grad::Bool=true)

Loads custom operator from CustomOps directory (shipped with ADCME instead of TensorFlow) For example

s = "SparseOperator"
oplib = "libSO"
grad = true

this will direct Julia to find library CustomOps/SparseOperator/libSO.dylib on MACOSX

ADCME.test_jacobian — Method

test_jacobian(f::Function, x0::Array{Float64}; scale::Float64 = 1.0)

Testing the gradients of a vector function f: y, J = f(x) where y is a vector output and J is the Jacobian.

ADCME.xavier_init — Function

xavier_init(size, dtype=Float64)

Returns a matrix of size size and its values are from Xavier initialization.

ADCME.compile — Method

compile(s::String)

Compiles the library s by force.

Misc